% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancers_drug_groups.R
\docType{data}
\name{cancers_drug_groups}
\alias{cancers_drug_groups}
\title{Data set 'cancers_drug_groups'}
\format{
An object of class \code{list} of length 2.
}
\usage{
cancers_drug_groups
}
\value{
No return value.
}
\description{
Data set 'cancers_drug_groups' is a list including a score dataframe with 147 drugs as rows and 19 cancer types as columns, and a dataframe with 9 self-defined drug groups (1st column) of the 147 drugs (2nd column). More details of the data can be found in the results of paper 'Tissue-specific identification of multi-omics features for pan-cancer drug response prediction' by Zhao Z., Wang S. Zucknick M. and Aittokallio T. (2022).
}
\examples{
# Load data
data(cancers_drug_groups, package = "EnrichIntersect")

}
\references{
Zhao Z, Wang S, Zucknick M and Aittokallio T (2019). \emph{Tissue-specific identification of multi-omics features for pan-cancer drug response prediction}. iScience, doi:10.1016/j.isci.2022.104767.
}
\keyword{datasets}
