% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancers_genes_drugs.R
\docType{data}
\name{cancers_genes_drugs}
\alias{cancers_genes_drugs}
\title{Data set 'cancers_genes_drugs'}
\format{
An object of class \code{array} of dimension 56 x 3 x 2.
}
\usage{
cancers_genes_drugs
}
\value{
No return value.
}
\description{
Data set 'cancers_genes_drugs' is an array with association scores between 56 genes (1st dimension), three cancer types (2nd dimension) and two drugs (3rd dimension). More details of the data can be found in the results of paper 'Tissue-specific identification of multi-omics features for pan-cancer drug response prediction' by Zhao Z., Wang S. Zucknick M. and Aittokallio T. (2022).
}
\examples{
# Load data
data(cancers_genes_drugs, package = "EnrichIntersect")

}
\references{
Zhao Z, Wang S, Zucknick M and Aittokallio T (2019). \emph{Tissue-specific identification of multi-omics features for pan-cancer drug response prediction}. iScience, doi:10.1016/j.isci.2022.104767.
}
\keyword{datasets}
