% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichIntersect-package.R
\docType{package}
\name{EnrichIntersect-package}
\alias{EnrichIntersect}
\alias{EnrichIntersect-package}
\title{EnrichIntersect: Enrichment Analysis and Intersecting Sankey Diagram}
\description{
A flexible tool for enrichment analysis based on user-defined sets. It allows users to perform over-representation analysis of the custom sets among any specified ranked feature list, hence making enrichment analysis applicable to various types of data from different scientific fields. 'EnrichIntersect' also enables an interactive means to visualize identified associations based on, for example, the mix-lasso model (Zhao et al., 2022 \doi{10.1016/j.isci.2022.104767}) or similar methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ocbe-uio/EnrichIntersect}
  \item Report bugs at \url{https://github.com/ocbe-uio/EnrichIntersect/issues}
}

}
\author{
\strong{Maintainer}: Zhi Zhao \email{zhi.zhao@medisin.uio.no}

Authors:
\itemize{
  \item Zhi Zhao \email{zhi.zhao@medisin.uio.no}
  \item Manuela Zucknick \email{manuela.zucknick@medisin.uio.no}
}

Other contributors:
\itemize{
  \item Tero Aittokallio \email{t.a.aittokallio@medisin.uio.no} [contributor]
}

}
\keyword{internal}
